Update graphql HTTP Response handling
The results of the https.request method are streamed to and resolved by the result.on("data") handler method, and while the current implementation will work with small amounts of data, if you have a large set of data it will only get a partial string and will fail to parse the JSON data throwing an error. 

This update combines the chunk data and then resolves the data with result.on("end") instead.